Clarify parser sandbox split
