+++ .vercelignore
@@ -1,0 +1,1 @@
+parser/sand
