+++ src/app/duckerweb/gh-flow-generator.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from "vitest";
 import fs from "node:fs";
-import { buildGhJson } from "parser/sand/src/parser";
+import { buildGhJson } from "parser/src/parser";
 import { generateFlowData } from "./gh-flow-generator";
 
 test("generateFlowData creates edges with correct structure", () => {
