+++ src/app/duckerweb/lib/component-handler.ts
@@ -1,4 +1,4 @@
-import type { Component } from "parser/sand/src/types";
+import type { Component } from "parser/src/types";
 import type { GHNode, GHNodeData, GHNodeType } from "../types/type";
 import type { NodePosition } from "./node-positions";
 import { getAccentColor } from "./node-classifier";
