+++ src/app/hooks/use-script-metrics.ts
@@ -1,8 +1,8 @@
 import { useState, useCallback } from "react";
 import { useFetchGhXml } from "./use-fetch-gh-xml";
-import { buildGhJson } from "parser/sand/src/parser";
+import { buildGhJson } from "parser/src/parser";
 import { generateFlowData } from "../duckerweb/gh-flow-generator";
-import type { ParsedGrasshopper } from "parser/sand/src/types";
+import type { ParsedGrasshopper } from "parser/src/types";
 import type { GHNode } from "../duckerweb/types/type";
 import type { Edge } from "@xyflow/react";
