+++ src/app/share/hooks/use-share-flow-state.ts
@@ -2,7 +2,7 @@ import { useAction } from "convex/react";
 import { api } from "@convex/_generated/api";
 import { useState, useEffect, useRef, useCallback } from "react";
 import { decompress } from "../../utils/gzip";
-import { buildGhJson } from "parser/sand/src/parser";
+import { buildGhJson } from "parser/src/parser";
 import { generateFlowData } from "../../duckerweb/gh-flow-generator";
 import type { GHNode } from "../../duckerweb/types/type";
 import type { Edge } from "@xyflow/react";
